backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
langchain
langchain-openai
langchain-core
langchain-community
pypdf
langchain-elasticsearch
langchain-ollama
langchain-text-splitters
langchain-huggingface
sentence-transformers
openai
structlog
httpx
tenacity
kafka-python>=2.0.2
cachetools>=5.3.0
